Fix: reset_number should be an integer and not a double.
git-svn-id: svn://svn.code.sf.net/p/sipp/code/sipp/trunk@307 60913f6b-29f6-41b8-8391-1ad2d2d6c344

+++ sipp.hpp
@@ -350,7 +350,7 @@ extern int           user_port                    _DEFVAL(0);
 extern char          hostname[80];
 extern bool          is_ipv6                      _DEFVAL(false);
 extern int           start_calls                  _DEFVAL(0);
-extern double        reset_number                 _DEFVAL(0);
+extern int           reset_number                 _DEFVAL(0);
 extern int	     reset_close                  _DEFVAL(1);
 extern int	     reset_sleep                  _DEFVAL(1000);
